Change "VMware" to "Omnissa"
The product name "VMware Identity Services" becomes "Omnissa Identity
Service".

The product name "VMware Identity Manager" is an obsolete synonym for
Omnissa Access, and it really shouldn't have been used in these
swagger docs ever. (It's referring to the wrong product.) Here, we just
update it to "Omnissa Identity Service", which is the correct product
name.

Also, this commit includes a few recent changes to the APIs, like adding
a new `auth_type_list` field to the directory APIs.

Signed-off-by: Joe Walbran <walbr037@morris.umn.edu>